Clean up build scripts
