+++ src/build-scripts/build-database/addArticles.ts
@@ -5,9 +5,9 @@ import {
   fromDocumentsRoot,
   stripFileExtensions,
   buildPathToRelatedItem
-} from './filePathHelpers';
+} from '../filePathHelpers';
 
-import { Article, Video, Collection } from '../models';
+import { Article, Video, Collection } from '../../models';
 
 import { ParsedArticle } from 'src/types/ParsedArticle';
 import { search } from 'src/controllers/searchController';
@@ -44,9 +44,7 @@ const addArticles = async (articles: ParsedArticle[]) => {
         }));
       }
       const relatedArticle = await Article.findOne({ where: { [searchKey]: searchValue } });
-      console.log('relatedArticle', relatedArticle);
       articleInstance.addRelatedArticle(relatedArticle);
-
     }
   }
 };
@@ -107,57 +105,5 @@ const getCollection = async (name: string) => {
     return collection;
   }
 };
-
-// const insertArticle = async (db, fileData) => {
-//   const { path, filePath, slug, html, ...otherFields } = fileData;
-//   const sql = `INSERT INTO articles(path, file_path, slug, body, data) VALUES (:path, :filePath, :slug, :body, :data)`;
-//   const result = await db.run(sql, {
-//     ':path': path,
-//     ':filePath': filePath,
-//     ':slug': slug,
-//     ':body': html,
-//     ':data': JSON.stringify(otherFields)
-//   });
-//   return result.lastID;
-// };
-
-// const associateArticleWithVideo = async (params) => {
-//   const { db, savedArticleId, articlePath, videoRelation } = params;
-//   const videoRelationPath = buildPathToRelatedItem({
-//     sourceFilePath: articlePath,
-//     relation: videoRelation
-//   });
-//   let { id: videoId } = await db.get(`SELECT id FROM videos WHERE file_path = "${videoRelationPath}"`) || {};
-//   if (! videoId) {
-//     const video = readVideo(videoRelationPath);
-//     videoId = await insertVideo(db, Object.assign(video, { filePath: videoRelationPath }));
-//   }
-
-//   const sql = 'INSERT INTO articles_videos(article_id, video_id) VALUES (:articleId, :videoId)';
-//   await db.run(sql, {
-//     ':articleId': savedArticleId,
-//     ':videoId': videoId
-//   });
-// };
-
-// const readVideo = (videoPath) => {
-//   const videoFileContent = fs.readFileSync(path.join(config.docsPath, videoPath), 'utf-8');
-//   return yaml.parse(videoFileContent);
-// };
-
-// const insertVideo = async (db, video) => {
-//   const { filePath, title, description, url, ...otherFields } = video;
-
-//   const sql = 'INSERT INTO videos(file_path, title, description, url, data) VALUES (:filePath, :title, :description, :url, :data)';
-//   const { lastID: id } = await db.run(sql, {
-//     ':filePath': filePath,
-//     ':title': title,
-//     ':description': description,
-//     ':url': url,
-//     ':data': JSON.stringify(otherFields)
-//   });
-
-//   return id;
-// };
 
 export default addArticles;
