+++ src/build-scripts/build-index/readSourceFiles.ts
@@ -3,7 +3,6 @@ import globby from 'globby';
 import config from '../../../config';
 import parseMarkdown from './parseMarkdown';
 import {
-  slugifyPath,
   stripFileExtensions
  } from '../filePathHelpers';
 
@@ -19,14 +18,12 @@ import {
   const articles = await Promise.all(filePaths.map(async (filePath) => {
     const pathWithinDocsDir = filePath.substring(docsPath.length + 1);
     const articlePath = stripFileExtensions(pathWithinDocsDir);
-    // let slug = slugifyPath(pathWithinDocsDir);
     const parsedFileContent = await parseMarkdown(filePath) as { text: string };
     return Object.assign(
       {},
       parsedFileContent,
       {
-        path: articlePath,
-        // slug
+        path: articlePath
       }
     );
   }));
